Timer + score & time saving

+++ SonicStarsAlign/Assets/SSA-Framework/Utility/Saving/StageData.cs
@@ -9,7 +9,7 @@ public class StageData : ScriptableObject
     public bool Complete = false;
     public byte Rank = 0;
     public List<bool> RedRings = new() { false, false, false, false, false };
-    public int bestTime = 0;
+    public float bestTime = 0;
     public int bestScore = 0;
     public List<int> RankScores = new() { 1000, 2000, 3000, 4000 };
     public int StageRingCount;
